# Backend/docs/swagger.yaml
definitions:
  controllers.AbsenRequest:
    properties:
      latitude:
        example: -7.250446
        type: number
      longitude:
        example: 112.768846
        type: number
      qr_token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      sesi_id:
        example: 1
        type: integer
    type: object
...
  controllers.ErrorMessage:
    properties:
      error:
        example: 'Terjadi kesalahan: data tidak valid'
        type: string
    type: object
...
  controllers.RiwayatAbsenSiswa:
    properties:
      mapel:
        example: Matematika
        type: string
      status:
        example: hadir
        type: string
      tanggal:
        example: "2026-04-08"
        type: string
    type: object
  controllers.SiswaData:
    properties:
      email:
        example: hafizh@example.com
        type: string
      id:
        example: 1
        type: integer
      nama_lengkap:
        example: Hafizh Junior
        type: string
      nis:
        example: "22010101"
        type: string
    type: object
...
  controllers.SuccessMessage:
    properties:
      message:
        example: Aksi berhasil dilakukan!
        type: string
    type: object
info:
  contact:
    email: hafizh@example.com
    name: Hafizh
  description: Ini adalah dokumentasi REST API untuk aplikasi absensi dengan fitur
    QR & Geofencing.
  termsOfService: http://swagger.io/terms/
  title: API Sistem Absensi Tempat Les
  version: "1.0"
paths:
...
  /admin/siswa/all:
    get:
      description: Mengambil daftar seluruh data master siswa
      produces:
      - application/json
      responses:
        "200":
          description: Data Siswa
          schema:
            items:
              $ref: '#/definitions/controllers.SiswaData'
            type: array
      security:
      - BearerAuth: []
      summary: Tampilkan Semua Data Siswa
      tags:
      - 5. Admin - Master Data
...
  /siswa/absen/submit:
    post:
      consumes:
      - application/json
      description: Mengirim token QR dinamis dan koordinat GPS Siswa untuk divalidasi
      parameters:
      - description: Kirim Koordinat dan Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.AbsenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Posisi tervalidasi, Absen Sukses
          schema:
            $ref: '#/definitions/controllers.SuccessMessage'
        "403":
          description: Di luar jangkauan radius / QR Kadaluwarsa
          schema:
            $ref: '#/definitions/controllers.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Submit Absensi Geofencing (Siswa)
      tags:
      - 2. Operasional Siswa
  /siswa/riwayat:
    get:
      description: Menampilkan data historis absensi milik siswa yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: Daftar Riwayat
          schema:
            items:
              $ref: '#/definitions/controllers.RiwayatAbsenSiswa'
            type: array
      security:
      - BearerAuth: []
      summary: Riwayat Absensi Siswa
      tags:
      - 2. Operasional Siswa
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"